pyproject.toml
[project]
name = "telegram-add-tool"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.12"
dependencies = ["aiogram (>=3.19.0,<4.0.0)", "fastapi (>=0.115.12,<0.116.0)", "uvicorn (>=0.34.0,<0.35.0)", "tortoise-orm (>=0.24.2,<0.25.0)", "aerich (>=0.8.2,<0.9.0)", "tomlkit (>=0.13.2,<0.14.0)", "asyncpg (>=0.30.0,<0.31.0)", "httpx (>=0.28.1,<0.29.0)", "python-dotenv (>=1.1.0,<2.0.0)", "black (>=25.1.0,<26.0.0)", "telethon (>=1.39.0,<2.0.0)", "typer (>=0.15.2,<0.16.0)"]

[[project.authors]]
name = "Овчинников Алексей Александрович"
email = "dev.allelleo@internet.ru"

[[tool.poetry.packages]]
include = "telegram_add_tool"
from = "src"

[tool.aerich]
tortoise_orm = "src.telegram_add_tool.backend.src.core.database.TORTOISE_CONFIG"
location = "./migrations"
src_folder = "./."

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
